enhence SolrClient::query.
add an optional argument to SolrClient::query method. the fragment will
be added to the end of the query url if its not null.

for example, with `$client->query($query, "Hello, World!");`, the last
part of the url will be `#Hello%2C%20World%21`.

this feature may be useful for tracing queries.